+++ packages/client/src/components/Routing/Routing.tsx
@@ -1,11 +1,12 @@
-import { Routes, Route, Link, RouteComponentProps } from 'react-router-dom'
-import { Login, Registration } from '../../pages'
+import { Routes, Route } from 'react-router-dom'
+import { Login, Registration, StartPage } from '../../pages'
 
 export const Routing = () => {
   return (
     <Routes>
-      <Route path="/registration" element={Registration} />
-      <Route path="/login" element={Login} />
+      <Route path="/registration" element={<Registration />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/" element={<StartPage />} />
     </Routes>
   )
 }
