+++ packages/client/src/pages/Login/Login.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const Login: FC = () => {
+  return (
+    <div className="registr">
+      <h1> Login</h1>
+    </div>
+  )
+}
