+++ packages/client/src/pages/StartPage/index.ts
@@ -1,0 +1,1 @@
+export { StartPage } from './StartPage'
