+++ packages/client/src/pages/index.ts
@@ -1,3 +1,4 @@
 import { Login } from './Login'
 import { Registration } from './Registration'
-export { Login, Registration }
+import { StartPage } from './StartPage'
+export { Login, Registration, StartPage }
